feat(model): add Hash trait to Arn struct
The Arn struct now derives the Hash trait. This allows instances of Arn to be used in hashed collections such as HashMap or HashSet. This change improves the flexibility and usability of the Arn struct within the Akton framework.

NEW FEATURE: You appear to have introduced one or more new features that are backward-compatible. According to Semantic Versioning (SemVer), this requires a minor version update. Please verify and update your version number accordingly.