chore(metadata): Update repository URLs and keywords
This commit updates the repository URLs and keywords in the Cargo.toml file. The repository URL is changed from 'https://github.com/govcraft/akton-arn' to 'https://github.com/govcraft/akton-eid'. The keywords are updated from 'arn' to 'eid'. The homepage URL is also updated to reflect the new repository name.