feat(traits): add Hash trait to Root
Add the Hash trait bound to the generic parameter T in the ErnComponent implementation for Root. This allows Root to be used in contexts where hashing is required.

NEW FEATURE: You appear to have introduced one or more new features that are backward-compatible. According to Semantic Versioning (SemVer), this requires a minor version update. Please verify and update your version number accordingly.